Revert "Move Zone (and subclasses) into Map-files"
This reverts commit c51ba1aad3c4837748eb85924dcb7cff1ed6fb4f.

Turns out that MapRenderers, which are in the views package, need to
access Zone classes.
Moving Zone into Map file made Zone classes package private in the
models package, as the only way to make a class public is to have it
in a file of the same name.
And thus, it's rewind time.