Mov all rng to RNG class, apply seeding
All random number generation has been moved to the RNG singleton class.
This uses a java.util.Random with a given seed, and any time this
object returns a new number it gives itself a new seed. This means that
all number generation is deterministic given the same starting seed.